requirements.txt
streamlit
python-dotenv
PyPDF2
threaded
google-generativeai